[Bug Fix] Changing iteration variable type to size_t in hash function

+++ Utilities.cpp
@@ -11,10 +11,10 @@ string hash(int key, const string& value1, const string& value2) {
     for (int i = 0; i < out_size; i++) {
         bytes[i] = key ^ (out_size - i);
     }
-    for (int i = 0; i < value1.size() * out_size ; i++) {
+    for (size_t i = 0; i < value1.size() * out_size ; i++) {
         bytes[i % (out_size/2)] ^= value1[i % value1.size()] ^ i;
     }
-    for (int i = 0; i < value2.size() * out_size ; i++) {
+    for (size_t i = 0; i < value2.size() * out_size ; i++) {
         bytes[(out_size/2) + (i % out_size/2)] ^= value2[i % value2.size()] ^ i;
     }
     for (int i = 0; i < out_size; i++) {
